refactor: extract 8 components + tehdit/risk reward system
COMPONENT EXTRACTION (src/popup/components/)
App.tsx 1042 → 664 satir (toplam 2000+ → 664, ~67% azaldi).
- StatusPanel: status panosu + konusma balonu + reasons + aksiyon
  butonlari (~400 satir)
- SettingsTab: Ayarlar sekmesinin tum JSX'i (3 SettingCard + whitelist
  + autofill chip + "Tum Ayarlar" linki)
- Header: marka butonu + Aktif toggle + bildirim cekmecesi butonu
- NotificationPanel: 🔔 paneli (welcome, protected days, daily summary,
  glossary)
- SettingCard: tek ayar karti (a11y button + CSS hover)
- ConfirmModal: 3 modal'in (Disable/Close/Trust) ortak shell'i
- DurumSkorCards: Durum sekmesindeki 3 sayac karti (Tarama / Tehdit
  / Potansiyel Risk) + filtreli listeler
- Footer: versiyon etiketi
- narrateReason: teknik reason → Alparslan agziyla anlatim helper'i

REWARD SYSTEM
Skor formulu artik ceza + odul model:
  - Tehdit > 0 → -10 puan/tehdit (eski)
  - Tehdit = 0 → +10 ODUL (yeni)
  - Risk > 0  → -5 puan/risk (eski)
  - Risk = 0  → +5 ODUL (yeni)

Panel rozetleri clean durumda da gosterilir (yesil "+10/+5 Puan"); i18n
metinleri sadelestirildi:
  - "Sistem analizi: 1 farkli tehdit vektoru..." → "1 adet tehlike engellendi."
  - "Tehdit vektoru saptanmadi, kalkanlar tam kapasitede." → "Tehdit bulunmadi"
  - "1 farkli potansiyel risk noktasi tespit edildi." → "1 adet potansiyel risk tespit edildi."
  - "Risk analizi temiz, anomali yok." → "Potansiyel risk bulunmadi"

Backend score-calculator + popup formula + 14 test (405/405 passing).
