+++ tests/dashboard/score-calculator.test.ts
@@ -4,48 +4,48 @@ import { calculateScore } from "@/dashboard/score-calculator";
 import { EMPTY_WEEKLY_METRICS, type WeeklyMetrics } from "@/dashboard/types";
 import t from "@/i18n/tr";
 
-describe("calculateScore (unique-domain penalty-based)", () => {
-  it("starts at 100 with no activity", () => {
+describe("calculateScore (unique-domain penalty + reward model)", () => {
+  it("starts at 100 with no activity (both clean rewards capped)", () => {
+    // Hicbir aktivite yok: +10 (tehdit yok) + 5 (risk yok) = +15 odul ama
+    // baslangic zaten 100 oldugu icin tavan 100'de kalir.
     const result = calculateScore(EMPTY_WEEKLY_METRICS);
     expect(result.score).toBe(100);
     expect(result.breakdown.threatAvoidanceScore).toBe(100);
     expect(result.tips).toContain(t.tips.notActive);
   });
 
-  it("deducts 10 points per DANGEROUS visit", () => {
+  it("1 DANGEROUS visit: -10 penalty + 5 risk-clean reward = 95", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 5,
       dangerousSitesVisited: 1,
     };
     const result = calculateScore(metrics);
-    expect(result.score).toBe(90);
+    expect(result.score).toBe(95);
     expect(result.tips.some((tip) => tip.includes("tehlikeli"))).toBe(true);
   });
 
-  it("deducts 10 points per SUSPICIOUS visit (threat bucket)", () => {
-    // Background tum tehditleri tek bir kovada (suspicious) toplar; 3
-    // benzersiz tehdit = -30.
+  it("3 SUSPICIOUS visits: -30 penalty + 5 risk-clean reward = 75", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 5,
       suspiciousSitesVisited: 3,
     };
     const result = calculateScore(metrics);
-    expect(result.score).toBe(100 - 30);
+    expect(result.score).toBe(75);
   });
 
-  it("deducts 5 points per UNKNOWN visit (risk bucket)", () => {
+  it("4 UNKNOWN visits: -20 penalty + 10 threat-clean reward = 90", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 10,
       unknownSitesVisited: 4,
     };
     const result = calculateScore(metrics);
-    expect(result.score).toBe(100 - 20);
+    expect(result.score).toBe(90);
   });
 
-  it("stacks penalties across categories", () => {
+  it("mixed activity: no rewards (both buckets have activity)", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 20,
@@ -53,6 +53,7 @@ describe("calculateScore (unique-domain penalty-based)", () => {
       suspiciousSitesVisited: 2, // -20
       unknownSitesVisited: 5,    // -25
     };
+    // Threats > 0 → no +10 reward; unknowns > 0 → no +5 reward
     const result = calculateScore(metrics);
     expect(result.score).toBe(100 - 10 - 20 - 25);
   });
@@ -61,76 +62,100 @@ describe("calculateScore (unique-domain penalty-based)", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 200,
-      // 20 tehdit * 10 = 200 > 100, floor 20
+      // 20 tehdit * 10 = 200 > 100; +5 risk-clean reward; floor 20
       suspiciousSitesVisited: 20,
     };
     const result = calculateScore(metrics);
     expect(result.score).toBe(20);
   });
 
-  it("returns 100 when only safe sites have been visited", () => {
+  it("perfect state with safe sites caps at 100", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 25,
       httpsCount: 25,
     };
-    const result = calculateScore(metrics);
+    const result = calculateScore(metrics, { uniqueSafeDomainCount: 5 });
     expect(result.score).toBe(100);
   });
 
-  it("subtracts 10 when 'Detayli Guvenlik Taramasi' (networkMonitoring) is off", () => {
-    // Ayar kapaliysa direkt -10 puan, UI'da kirmizi rozet ile gosterilir.
+  it("setting off does not negate rewards: -10 + 10 + 5 = 105 capped 100", () => {
     const result = calculateScore(EMPTY_WEEKLY_METRICS, {
       networkMonitoringEnabled: false,
     });
-    expect(result.score).toBe(90);
+    expect(result.score).toBe(100);
   });
 
-  it("does not add bonus when networkMonitoring is on", () => {
-    // Acikken ek puan EKLENMEZ, sadece -10 cezadan kacindirir. Skor
-    // maksimum 100'de kalir; UI yesil bilgi cumlesi gosterir.
+  it("setting on with no activity: rewards push but cap at 100", () => {
     const result = calculateScore(EMPTY_WEEKLY_METRICS, {
       networkMonitoringEnabled: true,
     });
     expect(result.score).toBe(100);
   });
 
-  it("combines visit penalties with settings penalty", () => {
+  it("3 suspicious + setting off: -30 -10 + 5 risk-clean = 65", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 10,
-      suspiciousSitesVisited: 3, // -30
+      suspiciousSitesVisited: 3,
     };
-    // -30 (threats) + -10 (setting off) = -40 -> 60
     const result = calculateScore(metrics, { networkMonitoringEnabled: false });
-    expect(result.score).toBe(60);
+    expect(result.score).toBe(65);
   });
 
-  it("adds 1 point per unique SAFE domain (capped at 100)", () => {
-    // 5 guvenli site bonusu = +5, ama baska ceza yok ve baslangic zaten
-    // 100 oldugu icin tavanda kapalı kaliyor.
-    const result = calculateScore(EMPTY_WEEKLY_METRICS, {
-      uniqueSafeDomainCount: 5,
-    });
-    expect(result.score).toBe(100);
-  });
-
-  it("uses safe bonus to offset risk penalties", () => {
+  it("3 suspicious + 15 safe: -30 + 15 + 5 risk-clean = 90", () => {
     const metrics: WeeklyMetrics = {
       ...EMPTY_WEEKLY_METRICS,
       urlsChecked: 30,
-      suspiciousSitesVisited: 3, // -30
+      suspiciousSitesVisited: 3,
     };
-    // -30 (threats) + +15 (15 safe * 1) = -15 puan -> 85
     const result = calculateScore(metrics, { uniqueSafeDomainCount: 15 });
-    expect(result.score).toBe(85);
+    expect(result.score).toBe(90);
   });
 
-  it("safe bonus cannot push score above 100", () => {
-    // 50 guvenli site, hicbir ceza yok → 100 + 50 = 150 ama cap 100.
+  it("safe bonus + rewards cannot push above 100", () => {
     const result = calculateScore(EMPTY_WEEKLY_METRICS, {
       uniqueSafeDomainCount: 50,
     });
     expect(result.score).toBe(100);
+  });
+
+  it("clean threat reward only applies when zero threats", () => {
+    // 1 dangerous = threats > 0 → no +10 reward; but risk clean → +5
+    const m1: WeeklyMetrics = {
+      ...EMPTY_WEEKLY_METRICS,
+      dangerousSitesVisited: 1,
+    };
+    expect(calculateScore(m1).score).toBe(100 - 10 + 5); // 95
+
+    // 1 suspicious also counts as threat
+    const m2: WeeklyMetrics = {
+      ...EMPTY_WEEKLY_METRICS,
+      suspiciousSitesVisited: 1,
+    };
+    expect(calculateScore(m2).score).toBe(100 - 10 + 5); // 95
+  });
+
+  it("clean risk reward only applies when zero unknowns", () => {
+    const m: WeeklyMetrics = {
+      ...EMPTY_WEEKLY_METRICS,
+      unknownSitesVisited: 1,
+    };
+    // 1 unknown = -5; no +5 reward; threat clean → +10 = 105 capped 100
+    expect(calculateScore(m).score).toBe(100);
+
+    // 2 unknowns => -10; threat clean +10 = 100; cap not hit
+    const m2: WeeklyMetrics = {
+      ...EMPTY_WEEKLY_METRICS,
+      unknownSitesVisited: 2,
+    };
+    expect(calculateScore(m2).score).toBe(100);
+
+    // 4 unknowns => -20; threat clean +10 = 90 (no cap)
+    const m3: WeeklyMetrics = {
+      ...EMPTY_WEEKLY_METRICS,
+      unknownSitesVisited: 4,
+    };
+    expect(calculateScore(m3).score).toBe(90);
   });
 });
